refactor: move unit tests for WalletProvider v1.0 and v1.3 implementations
